+++ R/load_relevant.R
@@ -24,8 +24,7 @@ load_relevant <- function(min_occurrences = 1000, min_species = 3) {
       by = "taxon_key"
     ) %>%
     transmute(
-      .data$year, .data$location, species = abbreviate(.data$scientific_name),
-      observed = 1L
+      .data$year, .data$location, species = .data$code, observed = 1L
     ) %>%
     pivot_wider(
       names_from = .data$species, values_from = .data$observed, values_fill = 0L