+++ NEWS.md
@@ -1,3 +1,14 @@
+# ladybird 0.0.2
+
+* Create functions:
+    * `base_model()`
+    * `secundary_model()`
+    * `cumulative_model()`
+* `import_data()` creates a code for the species based on their scientific name.
+* Add more Belgian data.
+* Add `run_analysis.R` to run the analyses.
+* Add `results.Rmd` to display the results.
+
 # ladybird 0.0.1
 
 * Create functions: